Made my branch better
