+++ apps/users/stats.py
@@ -1,28 +1,28 @@
-from django.db.models import Count
 import matplotlib.pyplot as plt
+from django.db.models import Count
 
 # Profile.objects.values('nationality').annotate(Count('nationality'))
 # dict = {nat[item]['nationality']:nat[item]['nationality__count'] for item in data]
 
 
-def getStats(query):
+def get_stats(query):
 
     # Creates a list of dictionaries of the nationalities and their count
-    departmentList = query.values("department").annotate(Count("department"))
-    programList = query.values("program").annotate(Count("program"))
-    nationalityList = query.values("nationality").annotate(Count("nationality"))
+    department_list = query.values("department").annotate(Count("department"))
+    program_list = query.values("program").annotate(Count("program"))
+    nationality_list = query.values("nationality").annotate(Count("nationality"))
 
     # Turn the list of dictionaries into one dictionary
-    department = {item["departement"]: item["department__count"] for item in departmentList}
-    program = {item["program"]: item["program__count"] for item in programList}
-    nationality = {item["nationality"]: item["nationality__count"] for item in nationalityList}
+    department = {item["department"]: item["department__count"] for item in department_list}
+    program = {item["program"]: item["program__count"] for item in program_list}
+    nationality = {item["nationality"]: item["nationality__count"] for item in nationality_list}
 
     # Call the plotting function (used in case the plotting function wants to be changed)
-    makePlot(department)
-    makePlot(program)
-    makePlot(nationality)
+    make_plot(department)
+    make_plot(program)
+    make_plot(nationality)
 
 
-def makePlot(data):
+def make_plot(data):
 
     plt.bar(data)
