+++ apps/users/templates/user/admin_add_stats_button.html
@@ -1,0 +1,12 @@
+{% extends 'admin/change_list.html' %}
+
+{% block object-tools %}
+    <div>
+        <form action="getstats/" method="POST">
+            {% csrf_token %}
+                <button type="submit">Make Immortal</button>
+        </form>
+    </div>
+    <br />
+    {{ block.super }}
+{% endblock %}
